backend/app/requirements.txt
fastapi
uvicorn[standard]
pillow
numpy
opencv-python
python-multipart
matplotlib
